services_no_otel/service-y/requirements.txt
flask
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
opentelemetry-sdk
opentelemetry-api
redis

